Add more warnings + small fixes

+++ src/colmatrix.h
@@ -7,7 +7,7 @@
 #include "svector.h"
 
 
-template <int m, int n = m>
+template <size_t m, size_t n = m>
 class ColMatrix {
   private:
     std::array<std::unique_ptr<SVector>, n> cols;
@@ -17,55 +17,55 @@ class ColMatrix {
         i = std::make_unique<SVector>();
     }
     ColMatrix(std::initializer_list<SVector>);
-    void add_value(int row, int col, double val) {
+    void add_value(size_t row, size_t col, double val) {
       cols[col]->add_value(row, val);
     }
-    const SVector& column(int i) const {
+    const SVector& column(size_t i) const {
       return *cols[i];
     }
-    void swapColumns(int i, int j);
+    void swapColumns(size_t i, size_t j);
     ColMatrix& operator=(ColMatrix& rhs) {
-      for(int i = 0; i < n; i++)
+      for(size_t i = 0; i < n; i++)
         cols[i] = std::make_unique<SVector>(rhs.column(i));
       return *this;
     }
     bool operator==(const ColMatrix& rhs) const;
     bool operator!=(const ColMatrix& rhs) const;
-    template <int mm, int nn>
+    template <size_t mm, size_t nn>
     friend std::ostream &operator<<(std::ostream &os, ColMatrix<mm,nn> const &matrix);
 };
 
-template <int m, int n>
+template <size_t m, size_t n>
 std::ostream &operator<<(std::ostream &os, ColMatrix<m,n> const &matrix) {
   for(auto& col : matrix.cols)
     os << *col << std::endl;
   return os;
 }
 
-template <int m, int n>
+template <size_t m, size_t n>
 ColMatrix<m,n>::ColMatrix(std::initializer_list<SVector> colList) {
   assert(colList.size() == n);
-  int i = 0;
+  size_t i = 0;
   for(auto& v : colList)
     cols[i++] = std::make_unique<SVector>(v);
 }
 
-template <int m, int n>
-void ColMatrix<m,n>::swapColumns(int i, int j) {
+template <size_t m, size_t n>
+void ColMatrix<m,n>::swapColumns(size_t i, size_t j) {
   assert(i >= 0 && i < m && j >= 0 && j < m);
   cols[i].swap(cols[j]);
 }
 
-template <int m, int n>
+template <size_t m, size_t n>
 bool ColMatrix<m,n>::operator==(const ColMatrix& rhs) const {
-  for(int i = 0; i < n; i++) {
+  for(size_t i = 0; i < n; i++) {
     if(column(i) != rhs.column(i))
       return false;
   }
   return true;
 }
 
-template <int m, int n>
+template <size_t m, size_t n>
 bool ColMatrix<m,n>::operator!=(const ColMatrix& rhs) const {
   return !(*this == rhs);
 }
